+++ src/sink.rs
@@ -447,4 +447,25 @@ mod tests {
             assert_eq!(queue_rx.next(), src.next());
         }
     }
+
+    #[test]
+    fn test_channel_volume() {
+        use crate::source::ChannelVolume;
+
+        let (sink, mut queue_rx) = Sink::new_idle();
+
+        let v = vec![100i16];
+        let input = SamplesBuffer::new(1, 48000, v.clone());
+
+        // High rate to avoid immediate control.
+        sink.append(ChannelVolume::new(input, vec![0.01, 0.01, 0.0, 0.0, 0.0, 0.0]));
+
+        assert_eq!(queue_rx.next(), Some(1.0));
+        assert_eq!(queue_rx.next(), Some(1.0));
+        assert_eq!(queue_rx.next(), Some(0.0));
+        assert_eq!(queue_rx.next(), Some(0.0));
+        assert_eq!(queue_rx.next(), Some(0.0));
+        assert_eq!(queue_rx.next(), Some(0.0));
+        assert_eq!(queue_rx.next(), None);
+    }
 }